Oh boy, make auo umlauts on long key press, which is trickier than initially thought.
thumb key now does triple duty as a WIN modifier, for the sole use of
mouse and window management in OpenBox. This is a bit confusing, falsely
triggers under Windows of course and is surprising as I can't use that
Win to press Win+F for, say Firefox shortcut. Will likely revisit once I
have a better thumb cluster.