Fix OSM keys to do what I mean.
This isn't fully debugged, but it's required to make rapid pressing of
layer → OSM CTL → t work to open a new tab, but also allow staying in
the layer so Ctrl+Cursor keys keep working.

I could also hold down ctrl on the base layer for that, then enter the
nav layer and press cursor keys there, but that seems too fragile.