Rework this to be just index based and not combine both readings.
Also fix the power up sequence, now that I got it working. Need to test
this at least on a Ploopy as well.

Also need to figure out how to merge the reports in my own code then.
I might also want to set the angle/rotation differently, so would need
to expose that as well.